requirements.txt
PyYAML==6.0.1
asgiref==3.7.2
attrs==23.2.0
Django==5.0.1
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-localflavor==4.0
drf-spectacular==0.27.2
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
mysqlclient==2.1.1
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3.post1
referencing==0.35.1
rpds-py==0.18.1
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
django-filter==24.2
